requirements.txt
streamlit
crewai
crewai-tools
langchain
langchain-community
langchain-core
python-dotenv
chromadb
pypdf
gTTS
unstructured
yt-dlp